PFMF-5146 - Budget chart in wrong color when just slightly above budget

+++ moneymanager-ui/src/main/java/com/tink/moneymanagerui/budgets/details/transactions/BudgetTransactionsListFragment.kt
@@ -88,7 +88,7 @@ internal class BudgetTransactionsListFragment : BaseFragment() {
             })
 
             amountLeft.observe(viewLifecycle, { amountLeft ->
-                budgetProgress.progress = if(amountLeft > 0) amountLeft else budgetProgress.max
+                budgetProgress.progress = if(amountLeft > 0) amountLeft.toInt() else budgetProgress.max
             })
 
             emptyState.observe(viewLifecycle, { isEmpty ->
